@@ -18,8 +18,8 @@ flutter:
         fileName: audioplayers_web.dart
 
 dependencies:
-  uuid: ^2.2.2
-  path_provider: ^1.6.22
+  uuid: ^3.0.4
+  path_provider: ^2.0.1
   flutter:
     sdk: flutter
   flutter_web_plugins:
@@ -28,9 +28,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  http: ^0.12.2
-  test: ^1.15.4
+  http: ^0.13.3
+  test: ^1.16.8
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0 <2.0.0"
