@@ -6,15 +6,15 @@ import 'package:flutter/services.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 class WrappedPlayer {
-  double pausedAt;
+  double? pausedAt;
   double currentVolume = 1.0;
   ReleaseMode currentReleaseMode = ReleaseMode.RELEASE;
-  String currentUrl;
+  String? currentUrl;
   bool isPlaying = false;
 
-  AudioElement player;
+  AudioElement? player;
 
-  void setUrl(String url) {
+  void setUrl(String? url) {
     currentUrl = url;
 
     stop();
@@ -34,8 +34,8 @@ class WrappedPlayer {
       return;
     }
     player = AudioElement(currentUrl);
-    player.loop = shouldLoop();
-    player.volume = currentVolume;
+    player!.loop = shouldLoop();
+    player!.volume = currentVolume;
   }
 
   bool shouldLoop() => currentReleaseMode == ReleaseMode.LOOP;
@@ -58,8 +58,8 @@ class WrappedPlayer {
     if (player == null) {
       recreateNode();
     }
-    player.play();
-    player.currentTime = position;
+    player!.play();
+    player!.currentTime = position;
   }
 
   void resume() {
@@ -67,7 +67,7 @@ class WrappedPlayer {
   }
 
   void pause() {
-    pausedAt = player.currentTime;
+    pausedAt = player!.currentTime as double?;
     _cancel();
   }
 
@@ -87,7 +87,7 @@ class WrappedPlayer {
 
 class AudioplayersPlugin {
   // players by playerId
-  Map<String, WrappedPlayer> players = {};
+  Map<String?, WrappedPlayer> players = {};
 
   static void registerWith(Registrar registrar) {
     final MethodChannel channel = MethodChannel(
@@ -100,11 +100,11 @@ class AudioplayersPlugin {
     channel.setMethodCallHandler(instance.handleMethodCall);
   }
 
-  WrappedPlayer getOrCreatePlayer(String playerId) {
+  WrappedPlayer getOrCreatePlayer(String? playerId) {
     return players.putIfAbsent(playerId, () => WrappedPlayer());
   }
 
-  Future<WrappedPlayer> setUrl(String playerId, String url) async {
+  Future<WrappedPlayer> setUrl(String? playerId, String? url) async {
     final WrappedPlayer player = getOrCreatePlayer(playerId);
 
     if (player.currentUrl == url) {
@@ -115,7 +115,7 @@ class AudioplayersPlugin {
     return player;
   }
 
-  ReleaseMode parseReleaseMode(String value) {
+  ReleaseMode parseReleaseMode(String? value) {
     return ReleaseMode.values.firstWhere((e) => e.toString() == value);
   }
 
@@ -125,13 +125,13 @@ class AudioplayersPlugin {
     switch (method) {
       case 'setUrl':
         {
-          final String url = call.arguments['url'];
+          final String? url = call.arguments['url'];
           await setUrl(playerId, url);
           return 1;
         }
       case 'play':
         {
-          final String url = call.arguments['url'];
+          final String? url = call.arguments['url'];
 
           // TODO(luan) think about isLocal (is it needed or not)
 
